key callback
getting old version with just keyCallBack added
