+++ test/java/yaw/TestHitBoxDifferentPlan.java
@@ -66,8 +66,8 @@ public class TestHitBoxDifferentPlan implements UpdateCallback {
         HitBox hb_2_1 =  cube_2_hitbox.fetchHitBox("hitbox 1");
         HitBox hb_2_2 =  cube_2_hitbox.fetchHitBox("hitbox 2");
 
-        if(hb_1_1.intersect(hb_2_1) || hb_1_1.intersect(hb_2_2) ||
-                hb_1_2.intersect(hb_2_1) || hb_1_2.intersect(hb_2_2)){
+        if(hb_1_1.isIsCollisionWith(hb_2_1) || hb_1_1.isIsCollisionWith(hb_2_2) ||
+                hb_1_2.isIsCollisionWith(hb_2_1) || hb_1_2.isIsCollisionWith(hb_2_2)){
             System.out.println("There is a collision");
             try {
                 Thread.sleep(3000);
